Create a ThemeOverlay for the legacy text field to unset editTextStyle.
The legacy text field had extra spacing from the editTextStyle set in the default textInputStyle. Since explicit styles stack on top of the default style, the editTextStyle with extra padding was set in the default style (which was recently changed to the filled box style) and it was not unset in the legacy style, resulting in the extra spacing.

Adding an empty ThemeOverlay fixes this issue.

PiperOrigin-RevId: 208117313