Add API to AppBarLayout Behavior to allow for disabling of offsetting
This makes it possible to disable the "scrolling" that happens when a scrolling
child has a height of match_parent but doesn't have enough actual content to
fill the screen (e.g., a RecyclerView with less than a few items). A scrolling
child with a height of wrap_content that doesn't fill the screen will already
not be scrolled/translated by the AppBarLayout Behavior.

PiperOrigin-RevId: 216255315