Use `colorOnSurface` to `colorSurface
My understanding of these attributes is that the "on" attributes (i.e. `colorOnPrimary`, `colorOnSurface`, etc.) are supposed to be used to describe text/icons when displayed on top of a background... and that the others (i.e. `colorPrimary`, `colorSurface`, etc.) are supposed to be used to describe the actual backgrounds.

If this is true, is it possible that there is a typo here and that you meant to use `colorSurface` instead?