Fix WindowInset handling in AppBarLayout
When AppBarLayout is used with a CollapsingToolbarLayout
child, it currently does not handle WindowInsets very
well. This commit fixes that by properly handling
insets.

The primary issue is when an AppBarLayout is used with a
single Toolbar (very common use case), where the Toolbar
will be displayed behind the status bar.

Another issue is with HeaderScrollingViewBehavior. To workaround
CoordinatorLayout using its 'compatible measuring' for insets, it
manually sets the scrolling view to fit system windows. This has
other issues though because by default that view will get padded in.
Fixed by manually adding the insets back into the measured dimensions.