Add a getter to check isExtended for ExtendedFloatingActionButton.
Context: This is useful to test the state of the extended fab post a shrink() or extend() call. Since we can't mock Views in unit tests, we can't verify calls on void methods for views, so this makes it such that we can assert the end state of those calls.
PiperOrigin-RevId: 242997029