Update from TextView to MaterialTextView
There are several places throughout the project that are still using the
TextView or the AppCompatTextView.  This causes issues with lineHeight not
being respected by multiple components in the project on devices that are
lower than API 28.

The changes here update all the components that are referencing the TextView or
the AppCompatTextView to use the MaterialTextView allow the lineheight to be
adjusted on all API versions supported by the library.