[ExtendedFloatingActionButton] Allowed for width to adapt to text size after expand animation
After the button has been expanded, the width is statically defined,
meaning it stays at a single value. If the text is changed, the width
will no longer grow and shrink with the size of the text, instead
creating whitespace and cutting out words. To fix it, the width is set
to WRAP_CONTENT after the animation happens, which allows the width
to change with the text.