Use text buttons for framework AlertDialogs in V21+. The library does not encourage usage of the framework AlertDialog (use the AppCompat one instead), however this change is to help support areas that do use the framework AlertDialog, like preferences.
Note: the default theme (android:alertDialogTheme) for framework alert dialogs pre-21 is not a public resource and therefore we cannot inherit from that theme. For this reason, we are not changing the buttons for AlertDialogs pre-21 (they will continue to use the framework default). They will, however, look like the typical pre-21 framework AlertDialogs once MaterialButton supports android:background.
PiperOrigin-RevId: 295263002