[Bottom Sheet] Updated peek height gesture inset behavior to only add extra inset if necessary
Previously on Q the peek height would have the bottom gesture inset size added to it regardless of whether edge to edge mode or gesture navigation were enabled. Now the extra inset will only be added if it is needed due to a system gesture conflict, in order to give the user some space to drag the sheet.

Resolves https://github.com/material-components/material-components-android/issues/1472

PiperOrigin-RevId: 320606747