Reset MaterialButtonToggleGroup children margins correctly
Background:
When changing a `MaterialButtonToggleGroup`'s orientation
programatically the children's margins are not reset correctly. This
causes the buttons to render incorrectly since the borders are missing.

Modifications:
* Update `MaterialButtonToggleGroup.adjustChildMarginsAndUpdateLayout`
to reset the top/start margins so the buttons render correctly.
* Add a new toggle to `ButtonToggleGroupDemoFragment` to programatically
switch the orientation of all the toggle groups between
vertical/horizontal in order to reproduce this issue.