[ShapeableImageView] Respect strokeWidth in ShapeMask
Problem:
When stroke is set it is rendered so that half of the stroke is inside of the bounds of the ImageView and half is outside. The problem is that anything that is outside of the bounds of the View gets clipped and thus won't be visible. That leads to incorrectly rendered stroke. Normally this should be fixed by telling the parent to not clip
its children, but that doesn't work here as here it gets actually clipped by the view itself by its ShapeMask.

Changes:
Offset the shape mask by half of the stroke width on all sides.

fixup! [ShapeableImageView] Respect strokeWidth in ShapeMask