[BottomNavigationView] Updated `BottomNavigationView` to draw compat shadows and elevation overlays in pre-21 and remove top divider for non-legacy backgrounds.
Previously, `BottomNavigationView` was using `MaterialShapeDrawable` but an elevation value that was set would not actually take effect. It turns out that this is because `BottomNavigationView` did not properly call the appropriate elevation methods. This commit fixes the handling of elevation. It also removes the top divider in pre-21 cases where there is a `MaterialShapeDrawable` background - otherwise the top divider is kept for compatibility with existing legacy versions.

PiperOrigin-RevId: 335867951