[Motion] Added basic music player and container transform to hero music player demo app.
Added a music player to the hero music player demo app, and a container transition from the FAB to the music player. The music player has a progress indicator and rewind, play, fast forward and volume controls; at this point, they all do nothing.

The transition uses a View to View container transform transition to transition from the FAB to the music player.

As of now, the transition does not touch the album view - it just jumps to underneath the music player for now.

PiperOrigin-RevId: 342915096