[Tab] Fix icon tint not applied on lower SDKs
On SDK < 23 if the icon is not a TintAwareDrawable, DrawableCompat.wrap() will return a new drawable which wraps the original drawable. However in the original logic after we set the tint to the icon, we didn't really set the tinted icon back to the tab or the image view, therefore the tint is not showing.

Moves the tint logic to updateTextAndIcon() to ensure tinted icons will be used in the image view.

Resolves https://github.com/material-components/material-components-android/issues/785

PiperOrigin-RevId: 391368322