[Menu] Fix toolbar menu arrow not visible
In toolbar styles we use android:theme to override text and icon colors and set ?attr/colorControlNormal to ?attr/colorOnPrimary, which is the same color (white) in the light mode as the menu background. To override this behavior again, we need to use ?attr/popupTheme and set ?attr/colorControlNormal to ?attr/colorOnSurface.

Resolves https://github.com/material-components/material-components-android/issues/553

PiperOrigin-RevId: 392003704